Refactor db_management.py to support different storage types and attach databases
This commit refactors the `db_management.py` file to support different storage types and attach databases accordingly. It introduces a new `attach_db` function that checks the `STORAGE_TYPE` environment variable and attaches the appropriate database. If the storage type is set to 'POSTGRES', it sets up a connection to a Postgres database and attaches it to DuckDB. Otherwise, it attaches a local DuckDB database using the `STORAGE_DIRECTORY` environment variable. The commit also includes a new `setup_postgre` function that installs and loads the Postgres plugin and sets up the connection to the Postgres database.